refactor: Enum 패키지명 수정
