+++ user-api/src/main/java/org/ecommerce/userapi/entity/enumerated/UserStatus.java
@@ -1,4 +1,4 @@
-package org.ecommerce.userapi.entity.type;
+package org.ecommerce.userapi.entity.enumerated;
 
 import org.ecommerce.common.utils.mapper.EnumMapperType;
 
@@ -7,10 +7,11 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public enum Role implements EnumMapperType {
+public enum UserStatus implements EnumMapperType {
 
-	USER("일반 유저"),
-	SELLER("판매 유저");
+	GENERAL("일반 회원"),
+	WITHDRAWAL("탈퇴 회원"),
+	SUSPEND("정지 회원");
 
 	private final String title;
 
@@ -18,4 +19,5 @@ public enum Role implements EnumMapperType {
 	public String getCode() {
 		return name();
 	}
+
 }
